+++ ChangeLog
@@ -1,7 +1,7 @@
 24/06/2020 Oleg Moskalenko <mom040267@gmail.com> Mihály Mészáros <misi@majd.eu>
 Version 4.5.2 'dan Eider':
 	- fix null pointer dereference in case of out of memory. (thanks to Thomas Moeller for the report)
-    - merge PR #517 (by wolmi)
+	- merge PR #517 (by wolmi)
 		* add prometheus metrics
 	- merge PR #637 (by David Florness)
 	    * Delete trailing whitespace in example configuration files
@@ -19,6 +19,8 @@ Version 4.5.2 'dan Eider':
 	- merge PR #599 (by Cédric Krier)
 		* Do not use FIPS and remove hardcode OPENSSL_VERSION_NUMBER with LibreSSL
 	- update Docker mongoDB and fix with workaround the missing systemctl
+	- merge PR #551 (by jelmd)
+		* support of --acme-redirect <URL>
 
 24/06/2020 Oleg Moskalenko <mom040267@gmail.com> Mihály Mészáros <misi@majd.eu>
 Version 4.5.1.3 'dan Eider':
